// src/app/layout/projectcompononts/models/menu-management.ts
export class MenuEntry {
    menuId: string
    menuName: string;
    menuDescription: string;
    menuUrl: string;
    menuicon: string ;
    orderNo: string;
    mainMenuId: string;
    createDate: string;
    isSubMenuCount: string;
    pageId:any;
    pageName:any;
    pageURL:any
}


export class Role{
    roleId:any;
    userId:any;
    orderNo:number;
    pageId:any;
    pageName:any;
    pageURL:any
} 